test/simple_test: Renamed from cmd/tuna
The previous 'tuna' command was only written to provide a basic test of the code
and is not a real application, therefore it is moved to the 'test' directory and
renamed.

The install rule for the 'cmd' directory is temporarily modified as it would
fail without any binaries to install.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>