bufq: Wait for queued data to be processed on exit
Instead of immediately signalling the consumer thread to exit when bufq_exit is
called, we enqueue an exit event and allow the consumer thread to finish
processing the data which has already been queued. This stops loss of data when
using a fast producer (eg input_sndfile) which may finish before the processing
and output code has finished.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>