libtuna: Ensure LDLIBRARIES_TGT is empty
By default, target-specific variables in Makefiles are recursive and apply to
all prerequisites built for that target. This causes a problem when
LDLIBRARIES_TGT specifies arguments to link against one of those prerequisites -
it cannot be linked against itself when building. So for libtuna we ensure that
LDLIBRARIES_TGT is empty.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>