fft: Disallow sharing of FFT contexts
Each time fft_set_length was called the FFT was being re-planned using the new
length. Thus only one length of FFT was supported at a time, despite the
interface acting as if multiple lengths were supported. The FFT would always be
performed at the longest set length and users which expected a shorter FFT would
end up with junk data included in their transform, ruining their entire output.

The FFT interface is therefore re-factored to disallow this sharing and all
users of the FFT interface are updated accordingly.

Whilst performing this refactor, the FFT interface is documented and a few
internal improvements are made to its operation.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>