bufhold: Fix return of bufhold_advance
Return values <0 are generally reserved to indicate errors. Therefore
bufhold_advance should not return -1 to indicate that a held buffer was released
as that is not an error condition. Instead the return value in this case is
changed to 0. This is consistant with the return value being the number of
samples remaining in the given held buffer.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>