tol: Provide band centre and band edge values
The 'freq' array is renamed to 'band_edges' to more accurately represent what it
contains and the values within it are given to a greater precision. A second
array, 'band_centres' is added.

These arrays may be accessed via the new 'tol_get_band_centre()' and
'tol_get_band_edge()' functions.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>