test/integration: Add zero to null test
This test uses the input_zero producer and the output_null consumer to do a
dummy data run. The counter consumer is used to end processing after 64k
samples.

This test is currently expected to fail as tuna erroneously returns '1' when
input_zero exits due to the stop flag being set.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>