time_slice: Do not process half a time slice
We should only call process_time_slice() when we have enough data for a full
time slice, knowing that the time slice length is twice the time slice period
due to the 50% overlap. As we only reduce the amount of data available by the
time slice period (not the time slice length) when processing occurs, this will
still happen once per time slice period.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>