env_estimate: Use sample magnitude not energy
This change simplifies things in several places.

The sample_limit argument to env_estimate_init() is no longer required and so it
is removed. In turn, it is no longer needed in struct pulse_params and is again
removed.

As threshold_limit is now in terms of magnitude not energy, the value of 16 is
replaced by 4.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>