onset_threshold: Merge in minima tracker
The only user of the minima tracker was onset_threshold and so in order to
optimise the critical path in the pulse detection code minima is merged into
onset_threshold and no longer exists as its own module.

The new onset_threshold_next function takes a pointer to a threshold value which
is only updated when the underlying minimum is updated. This has a measurable
impact on performance!

The function onset_threshold_current is added to allow access to the current
threshold value.

The function onset_threshold_current_minimum is added to allow access to the
current mimimum as would have been reported by minima_current. This function is
used to test that the underlying minima tracking is still working correctly.

Several other files are updated due to these changes.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>