analysis: Fix segfault on pulse_init() failure
If pulse_init() failed, the error handler was called before a->time_slice was
given a value. Thus its value was non-NULL junk and consumer_exit() was called
on this junk. Inevitably this caused a segfault.

This is fixed by using calloc() instead of malloc(), thus ensuring that
a->time_slice is NULL before consumer_new() has been called for it.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>