offset_threshold: Fix and improve delay line usage
The delay line was previously filled with zeros each time offset_threshold_reset
was called. This was an expensive operation and was actually incorrect!

Instead we maintain a count of how many samples we have seen since the last call
to offset_threshold_reset. If this count is less than the delay line length then
the output of the delay line is junk and won't be used.

Instead of putting sample values into the delay line and tracking a delayed
minimum from the output of the delay line, we put minima into the delay line and
simply get the delayed minima out.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>