swig: Clean up exposure of logging functions
The logging wrappers provided by libtuna are intended for use in native C code
and probably won't work so well in other languages due to the differences in
varargs handling. So we don't provide bindings for these functions. Instead a
simple log_print() binding is given which logs a simple string.

The unit test for the logging functions is updated accordingly.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>