tol: Fix incorrect number of levels
The number of third octave levels was calculated based on the sample rate which
caused problems when the analysis length was not equal to the sample rate.
Instead it should be based on the number of frequency domain samples which are
processed, which is equal to half the analysis length.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>