time_slice: Avoid possible uint wrap-around in compare
When looping around vectorised operations we need to ensure there are at least 4
elements remaining in the array to be processed. If we do this by subtracting 3
from the number of elements in the array we will run into problems if that
number is less than three. Instead we should add 3 to the index variable.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>