pulse: Fix unsigned integer subtraction
If both onset_threshold_age(...) and count are unsigned integers, the expression
`count - onset_threshold_age(...)` will always be positive, even if count is
smaller than onset_threshold_age(...). By doing a comparison first and then
calculating the correct expression we can correctly identify the case where
count is smaller.

If count is larger than the onset threshold age, we still need to call
discard_leading_data but with zero as the offset.

Signed-off-by: Paul Barker <paul@paulbarker.me.uk>